+++ Include/Acidanthera/Library/OcAfterBootCompatLib.h
@@ -15,6 +15,8 @@
 #ifndef OC_AFTER_BOOT_COMPAT_LIB_H
 #define OC_AFTER_BOOT_COMPAT_LIB_H
 
+#include <Library/OcCpuLib.h>
+
 /**
   Booter patch structure.
 **/
@@ -196,12 +198,14 @@ typedef struct OC_ABC_SETTINGS_ {
   incompatible firmware to prevent boot failure and UEFI services breakage.
 
   @param[in]  Settings  Compatibility layer configuration.
+  @param[in]  CpuInfo   CPU information.
 
   @retval EFI_SUCCESS on success.
 **/
 EFI_STATUS
 OcAbcInitialize (
-  IN OC_ABC_SETTINGS  *Settings
+  IN OC_ABC_SETTINGS  *Settings,
+  IN OC_CPU_INFO      *CpuInfo
   );
 
 #endif // OC_AFTER_BOOT_COMPAT_LIB_H