+++ Library/OcMainLib/OpenCoreUefi.c
@@ -511,7 +511,8 @@ OcInstallPermissiveSecurityPolicy (
 
 VOID
 OcLoadBooterUefiSupport (
-  IN OC_GLOBAL_CONFIG  *Config
+  IN OC_GLOBAL_CONFIG  *Config,
+  IN OC_CPU_INFO       *CpuInfo
   )
 {
   OC_ABC_SETTINGS        AbcSettings;
@@ -656,7 +657,7 @@ OcLoadBooterUefiSupport (
   AbcSettings.ExitBootServicesHandlers = mOcExitBootServicesHandlers;
   AbcSettings.ExitBootServicesHandlerContexts = mOcExitBootServicesContexts;
 
-  OcAbcInitialize (&AbcSettings);
+  OcAbcInitialize (&AbcSettings, CpuInfo);
 }
 
 VOID
@@ -728,7 +729,7 @@ OcLoadUefiSupport (
   //
   // Setup Apple bootloader specific UEFI features.
   //
-  OcLoadBooterUefiSupport (Config);
+  OcLoadBooterUefiSupport (Config, CpuInfo);
 
   if (Config->Uefi.Quirks.IgnoreInvalidFlexRatio) {
     OcCpuCorrectFlexRatio (CpuInfo);